requirements.txt.txt
flask
gunicorn
numpy
tensorflow
pillow
opencv-python